1) Cleaned up code/indentation for most files.
2) BGGraphWindow was used specifically for NAMD analysis a while back and
is removed to avoid clutter.

3) Timeline tools have acquired a progress bar when data is being read. It
is not perfect (there is a inexplicable pause at the end of the data read),
but will do for now until the code gets refactored into a cleaner form.

4) Major refactoring done on Analysis.java. StsReader is now effectively
isolated from the rest of the code, accessible via Analysis alone. The
reason for this is that I intend for Analysis (which is a static class) to
eventually become the template for the abstraction of a ProjectionsRun, each
of which will play host to an StsReader. Analysis.java will also host an
IntervalData object (internally compressed) and a PointData object (not
yet implemented) representing the two main data abstractions used in
projections thus far. Only "modern" code uses IntervalData to acquire its
data. Legacy codes still read directly from the various Readers via Analysis
and the intention of further refactoring is to eliminate this.

5) FloatTextField has been updated to return the last kept value instead of
a hardcoded 1.0 in the event that some invalid floating point number was
entered directly into the text field.

6) Stl (Overview) tools have been modified with the intention of supporting
entry point colors. Unfortunately, this functionality rests on the ability
of Overview to use IntervalData instead of its current direct access to
LogReader through LoadGraphData. The old functionality is intact, but
attempts to use the "entry point" radio button in the newly designed window
will fail.

7) ProfileWindow now makes use of the progress bar correctly.

8) All other changes involve the refactoring of Analysis where access to
StsReader now goes through Analysis instead of direct access.