1) The code has been cleaned up and indentation fixed.
2) Fixed the bug where the scroll bars refuse to react properly to a Graph
canvas client resize. Scaling of the graph still does not work as it should
and will not cause the scroll bars to react properly. Work is currently being
done to support a "fit-to-screen" mode where data is either merged (re-binned)
or "thrown away" (squeezed into one pixel) when large amounts of data is
required to fit into a small space.

3) Default color acquisition schemes in DataSource has been fixed to supply
colors when a color map is available and the red/green/blue/gray cycle when
either the map is not present or when not enough color space has been
assigned (actually a bug) for the number of y values supplied.