Added a new feature to the Projections windowing and dialog framework that enforces consistency between the parameters maintained by the windows and those maintained by the dialog boxes. It also allows the dialog boxes to be synchronized with new accepted parameters set by the window (without the dialog's knowledge).
Added a new dialog type called SimpleThresholdDialog. It's job is to maintain
a simple threshold field (double) that is used by the tool (window) to filter
data failing to meet that one threshold criteria.

The only employing window - HistogramWindow recognizes the threshold but has
not been modified to actually make use of it for filtering purposes.