Refactored how colors are saved and initialized in Projections.
This will now be activity-based with one array for each activity. Much of
color management code has now been extracted from the Analysis class and
moved to a ColorManager class which is now responsible for supplying
facilities and colors to Analysis and/or Analysis-type Objects.

The follow up to this should be to implement a unified Color selection
dialog window for all tools that will work on the supplied Color array
acquired by Analysis.