Added the ability to add lines to timeline from the outlier analysis tool.
This involved an infrastructural change to gui/graph/Graph, allowing it
to respond to Mouse click events and to propagate it to tools that are
written to respond to such events.

This is a functional prototype in that it will not break the rest of
projections but it is not perfect. It is noted that the addProcessor() call
actually reloads all timelines already previously loaded as Sanjay suspected
so this will have to be reworked somehow.