Use relative time to reduce precision of timestamp fields
Instead of absolutely storing every timestamp as a long, store only
the relevant beginning timestamps absolutely as longs, storing
everything else as relative ints. Note that this has the limitation
that the relative timestamps for a particular EntryMethodObject have
to be within 2^31 microseconds (or ~35.8 minutes) of the absolute
timestamp, an assumption that will likely never be violated.