Compress EP ID data
Convert EP ID to a short, as it is on the C++ tracing side and remove
mostly superfluous entryIndex. This also required creating flags for
idle and overheap EMOs, they were previously stored as -1 and -2 in
the EP ID, which is now being treated as an unsigned short; all flags
are now being packed into a byte, so they have no additional memory
footprint.