Fix ending position of EMO blocks in timeline
The ending coordinate of an EMO should be on the pixel before the
pixel corresponding to its ending time (in other words, the beginning
should be inclusive, the ending should be exclusive). Without this, a
block that starts at the same time as the previous one ends would be
misleadingly shown to overlap.